+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+## [Unreleased]
+
+* Manual Tracing for lambda.
+* Configurable withOpenTelemetry Hook for Lambda
 
 ## [0.5.5] - 2024-03-28